requirements.txt
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.3.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
